Remove backward for im2col and col2im
`im2col` is a linear map, and `col2im` is its adjoint. As such, the
adjoint to `col2im` is `im2col` (the adjoint of the adjoint is the
original function.

There's no point having explicit derivatives in ATen for these
functions, so this PR deletes all these.

Furthermore, along the way, we fix an error for the derivative of im2col
for non-batched inputs.

ghstack-source-id: 416a3a32a54de4491a2d9dd4e6f371770348ee4d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85542